tests(helpers): move client functions
